+++ resources/tag.py
@@ -1,13 +1,13 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from models import TagModel, StoreModel, ItemModel
 from schemas import TagSchema, TagAndItemSchema
 
 blp = Blueprint("Tags", "tags", description="Operations on tags")
-
 
 @blp.route("/store/<int:store_id>/tag")
 class TagsInStore(MethodView):
@@ -17,9 +17,16 @@ class TagsInStore(MethodView):
 
         return store.tags.all()
 
+    # TODO: Add description to the blp response 201 object
+    @jwt_required()
     @blp.arguments(TagSchema)
     @blp.response(201, TagSchema)
     def post(self, tag_data, store_id):
+
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privileges required.")
+
         tag = TagModel(**tag_data, store_id=store_id)
 
         try:
@@ -33,8 +40,16 @@ class TagsInStore(MethodView):
 
 @blp.route("/item/<int:item_id>/tag/<int:tag_id>")
 class LinkTagsToItem(MethodView):
+
+    # TODO: Add description to the blp response 201 object
+    @jwt_required()
     @blp.response(201, TagSchema)
     def post(self, item_id: int, tag_id: int):
+
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privileges required.")
+
         item = ItemModel.query.get_or_404(item_id)
         tag = TagModel.query.get_or_404(tag_id)
 
@@ -45,43 +60,57 @@ class LinkTagsToItem(MethodView):
             db.session.commit()
         except SQLAlchemyError:
             abort(500, message="An error occurred while inserting the tag.")
-  
+
         return tag
-    
-    
+
+    # TODO: Add description to the blp response 200 object
+    @jwt_required()
     @blp.response(200, TagAndItemSchema)
     def delete(self, item_id, tag_id):
+
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privileges required.")
+
         item = ItemModel.query.get_or_404(item_id)
         tag = TagModel.query.get_or_404(tag_id)
-        
         item.tags.remove(tag)
-        
+
         try:
             db.session.add(item)
             db.session.commit()
         except SQLAlchemyError:
             abort(500, message="An error occurred while removing the tag.")
-            
+
         return {"message": "Item removed from tag", "item": item, "tag": tag}
 
 @blp.route("/tag/<int:tag_id>")
 class Tag(MethodView):
+    # TODO: Add description to 200 response code annotation
     @blp.response(200, TagSchema)
     def get(self, tag_id: int):
         tag = TagModel.query.get_or_404(tag_id)
         return tag
-    
-    
-    @blp.response(202, description="Deletes a tag if no item is tagged with it.", example={"message": "Tag deleted."})
+
+    @jwt_required()
+    @blp.response(202, description="Deletes a tag if no item is tagged with it.",
+                  example={"message": "Tag deleted."})
     @blp.alt_response(404, description="Tag not found.")
-    @blp.alt_response(400, description="Returned if the tag is assigned to one or more items. In this case, the tag is not deleted.")
+    @blp.alt_response(400, description="Returned if the tag is assigned to one or more items." +
+                    "In this case, the tag is not deleted.")
     def delete(self, tag_id: int):
+
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privileges required.")
+
         tag = TagModel.query.get_or_404(tag_id)
-        
+
         # Checks if the items list associated with this tag is empty
         if not tag.items:
             db.session.delete(tag)
             db.session.commit()
             return {"message": "Tag deleted."}
-        
-        abort(400, message="Could not delete tag. Make sure tag is not associated with any items, then try again.")
+
+        abort(400, message="Could not delete tag. Make sure tag is not associated" +
+            "with any items, then try again.")
